% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperbolicDEA.R
\name{hyperbolicDEA}
\alias{hyperbolicDEA}
\title{hyperbolicDEA}
\usage{
hyperbolicDEA(
  X,
  Y,
  RTS = "vrs",
  WR = NULL,
  SLACK = F,
  ACCURACY = 1e-10,
  XREF = NULL,
  YREF = NULL,
  SUPEREFF = F,
  NONDISC_IN = NULL,
  NONDISC_OUT = NULL,
  PARALLEL = 1
)
}
\arguments{
\item{X}{Matrix or dataframe with DMUS as rows and inputs as columns}

\item{Y}{Matrix or dataframe with DMUs as rows and outputs as columns}

\item{RTS}{Character string indicating the returns-to-scale, e.g. "crs", "vrs", "ndrs", "nirs", "fdh"}

\item{WR}{Matrix with one row per homogenous linear weight restriction in standard form, ncol(WR) = ncol(X) + ncol(Y)}

\item{SLACK}{Variable indicating whether an additional estimation of slacks shall be performed}

\item{ACCURACY}{Accuracy value for non-linear programmer}

\item{XREF}{Matrix or dataframe with firms defining the technology as rows and inputs as columns}

\item{YREF}{Matrix or dataframe with firms defining the technology as rows and outputs as columns}

\item{SUPEREFF}{Variable indicating whether superefficiencies shall be estimated}

\item{NONDISC_IN}{Vector containing indices of the input matrix that are non-discretionary variables}

\item{NONDISC_OUT}{Vector containing indices of the output matrix that are non-discretionary variables}

\item{PARALLEL}{Integer of amount of cores that should be used for estimation (Check availability of computer)}
}
\value{
A list object containing efficiency scores, lambdas, and potentially slacks and
binding parameters in the weight restrictions (mus)
}
\description{
Hyperbolic DEA implementation including weight restrictions,
exogenous variables, external referencing, estimation of slacks and
superefficiency scores
}
\examples{
X <- c(1,1,2,4,1.5,2,4,3)
Y <- c(1,2,4,4,0.5,2.5,3.5,4)
hyperbolicDEA(X,Y,RTS="vrs", SUPEREFF = F)
}
